Code clean

+++ orion-viewer/src/main/java/universe/constellation/orion/viewer/search/SearchTask.java
@@ -121,7 +121,7 @@ public abstract class SearchTask {
 
             @Override
             protected void onProgressUpdate(Integer... values) {
-                progressDialog.setProgress(values[0].intValue());
+                progressDialog.setProgress(values[0]);
             }
 
             @Override
